+++ aDVanceERP.Modulos.Inventario/MVP/Vistas/Movimiento/Plantillas/IVistaRegistroMovimiento.cs
@@ -7,7 +7,7 @@ public interface IVistaRegistroMovimiento : IVistaRegistro {
     string? NombreAlmacenOrigen { get; set; }
     string? NombreAlmacenDestino { get; set; }
     DateTime Fecha { get; set; }
-    int CantidadMovida { get; set; }
+    float CantidadMovida { get; set; }
     string TipoMovimiento { get; set; }
