+++ aDVancePOS.Modulos.TerminalVenta/MVP/Vistas/Venta/Plantillas/IVistaModificadorCantidadProducto.cs
@@ -3,5 +3,5 @@
 namespace aDVancePOS.Modulos.TerminalVenta.MVP.Vistas.Venta.Plantillas; 
 
 public interface IVistaModificadorCantidadProducto : IVista {
-    int CantidadProducto { get; set; }
+    float CantidadProducto { get; set; }
 }
